adding display info to search response data
