+++ learning_resources_search/serializers.py
@@ -36,7 +36,9 @@ from learning_resources.models import (
 from learning_resources.serializers import (
     ContentFileSerializer,
     CourseNumberSerializer,
+    LearningResourceMetadataDisplaySerializer,
     LearningResourceSerializer,
+    LearningResourceWithDisplayInfoSerializer,
     MicroLearningPathRelationshipSerializer,
     MicroUserListRelationshipSerializer,
 )
@@ -672,6 +674,13 @@ class LearningResourcesSearchResponseSerializer(SearchResponseSerializer):
                         ).data
                     )
 
+    def add_display_data(self, hits):
+        for hit in hits:
+            if hit.get("_source"):
+                hit["_source"]["display_data"] = (
+                    LearningResourceMetadataDisplaySerializer(hit["_source"]).data
+                )
+
     def update_list_parents(self, hits, user):
         """Fill in user_list_parents for users"""
         user_list_parents_dict = {}
@@ -694,10 +703,11 @@ class LearningResourcesSearchResponseSerializer(SearchResponseSerializer):
                     ).data
                 )
 
-    @extend_schema_field(LearningResourceSerializer(many=True))
+    @extend_schema_field(LearningResourceWithDisplayInfoSerializer(many=True))
     def get_results(self, instance):
         hits = instance.get("hits", {}).get("hits", [])
         request = self.context.get("request")
+        self.add_display_data(hits)
         if request and request.user and request.user.is_authenticated:
             self.update_list_parents(hits, request.user)
             if (
