+++ vector_search/serializers.py
@@ -1,11 +1,5 @@
-import datetime
-import json
-
-from dateutil import parser as date_parser
-from django.conf import settings
 from drf_spectacular.plumbing import build_choice_description_list
 from drf_spectacular.utils import extend_schema_field
-from langchain_text_splitters import RecursiveJsonSplitter
 from rest_framework import serializers
 
 from learning_resources.constants import (
@@ -26,197 +20,6 @@ from learning_resources_search.serializers import (
     SearchResponseMetadata,
     SearchResponseSerializer,
 )
-
-
-class LearningResourceMetadataDisplaySerializer(serializers.Serializer):
-    """
-    Serializer to render course information as a text document
-    """
-
-    title = serializers.CharField(read_only=True)
-    description = serializers.CharField(read_only=True)
-    platform = serializers.ReadOnlyField(read_only=True)
-    offered_by = serializers.ReadOnlyField(read_only=True)
-    full_description = serializers.CharField(read_only=True)
-    url = serializers.CharField(read_only=True)
-    departments = serializers.ReadOnlyField(
-        read_only=True,
-    )
-    free = serializers.ReadOnlyField(read_only=True)
-    topics_display = serializers.SerializerMethodField()
-    price_display = serializers.SerializerMethodField()
-    certification_display = serializers.SerializerMethodField()
-    instructors_display = serializers.SerializerMethodField()
-    runs_display = serializers.SerializerMethodField()
-    offered_by_display = serializers.SerializerMethodField()
-    languages_display = serializers.SerializerMethodField()
-    levels_display = serializers.SerializerMethodField()
-    departments_display = serializers.SerializerMethodField()
-    platform_display = serializers.SerializerMethodField()
-
-    def get_departments_display(self, serialized_resource):
-        department_vals = []
-        for department in serialized_resource.get("departments", []):
-            school_display = (
-                f" ({department['school']['name']})" if department.get("school") else ""
-            )
-            department_vals.append(f"{department['name']}{school_display}")
-        return ", ".join(department_vals)
-
-    def get_platform_display(self, serialized_resource):
-        return (
-            serialized_resource["platform"].get("name")
-            if serialized_resource.get("platform")
-            else ""
-        )
-
-    def get_levels_display(self, serialized_resource):
-        levels = []
-        for run in serialized_resource.get("runs", []):
-            if run.get("level"):
-                levels.extend(lvl["name"] for lvl in run["level"])
-        return ", ".join(set(levels))
-
-    def get_languages_display(self, serialized_resource):
-        languages = []
-        for run in serialized_resource.get("runs", []):
-            if run.get("languages"):
-                languages.extend(run["languages"])
-        return ", ".join(set(languages))
-
-    def get_offered_by_display(self, serialized_resource):
-        return (
-            serialized_resource["offered_by"].get("name")
-            if serialized_resource.get("offered_by")
-            else ""
-        )
-
-    def get_runs_display(self, serialized_resource):
-        runs = []
-        for run in serialized_resource.get("runs", []):
-            start_date = run.get("start_date")
-            formatted_date = (
-                date_parser.parse(start_date)
-                .replace(tzinfo=datetime.UTC)
-                .strftime("%B %d, %Y")
-                if start_date
-                else ""
-            )
-            location = run.get("location") or "Online"
-            duration = run.get("duration")
-            delivery_modes = (
-                ", ".join(delivery["name"] for delivery in run.get("delivery", []))
-                or "Not specified"
-            )
-            instructors = ", ".join(
-                instructor["full_name"]
-                for instructor in run.get("instructors", [])
-                if "full_name" in instructor
-            )
-            runs.append(
-                f" - Start Date: {formatted_date}, Location: {location}, "
-                f"Duration: {duration}, Format: {delivery_modes},"
-                f" Instructors: {instructors}"
-            )
-        return "\n".join(runs) if runs else ""
-
-    def get_instructors_display(self, serialized_resource):
-        instructors = set()
-        for run in serialized_resource.get("runs", []):
-            for instructor in run.get("instructors", []):
-                instructors.add(instructor["full_name"])
-
-        return ", ".join(instructors)
-
-    def get_certification_display(self, serialized_resource):
-        if serialized_resource.get(
-            "certification_type"
-        ) and not serialized_resource.get("free"):
-            return serialized_resource["certification_type"].get("name")
-        if (
-            serialized_resource["free"]
-            and serialized_resource["certification_type"]
-            and serialized_resource["certification_type"]["code"]
-            == CertificationType.none.name
-        ):
-            return CertificationType.none.value
-        return None
-
-    def get_price_display(self, serialized_resource):
-        prices = serialized_resource.get("prices", [])
-        if not serialized_resource["free"] and prices:
-            return f"${serialized_resource['prices'][0]}"
-        if (
-            serialized_resource["free"]
-            and serialized_resource["certification_type"]
-            and serialized_resource["certification_type"]["code"]
-            == CertificationType.completion.name
-        ) and len(prices) > 1:
-            return f"Free (Earn a certificate: ${prices[1]})"
-        return "Free"
-
-    def get_topics_display(self, serialized_resource):
-        return ", ".join(
-            topic["name"] for topic in serialized_resource.get("topics", [])
-        )
-
-    def render_document(self):
-        data = self.data
-        display_sections = {
-            "title": "Title",
-            "platform_display": "Platform",
-            "url": "Link",
-            "delivery": "Format",
-            "departments_display": "Departments",
-            "description": "Description",
-            "full_description": "Full Description",
-            "topics_display": "Topics",
-            "price_display": "Cost",
-            "certification_display": "Certification",
-            "instructors_display": "Instructors",
-            "runs_display": "Course Schedule (Runs and Sessions)",
-            "offered_by_display": "Offered By",
-            "languages_display": "Languages",
-            "levels_display": "Levels",
-        }
-        rendered_data = {}
-
-        for section, section_display in display_sections.items():
-            display_text = data.get(section)
-            if display_text:
-                rendered_data[section_display] = display_text
-        return rendered_data
-
-    def _json_to_text_document(
-        self, json_text, document_prefix="Information about this course:"
-    ):
-        """Render a (serialized) json fragment as plain text"""
-        rendered_info = f"{document_prefix}\n"
-        for section, section_value in json.loads(json_text).items():
-            rendered_info += f"{section} -\n{section_value}\n"
-        return rendered_info
-
-    def render_chunks(self):
-        rendered_doc = self.render_document()
-        """
-        We cant use tiktoken for token size calculation so
-        we use a rough calculation of 4*chunk_size characters:
-        https://help.openai.com/en/articles/4936856-what-are-tokens-and-how-to-count-them
-        """
-        chunk_size = (
-            settings.CONTENT_FILE_EMBEDDING_CHUNK_SIZE_OVERRIDE
-            if settings.CONTENT_FILE_EMBEDDING_CHUNK_SIZE_OVERRIDE
-            else 512
-        )
-        return [
-            self._json_to_text_document(json_fragment)
-            for json_fragment in RecursiveJsonSplitter(
-                max_chunk_size=chunk_size * 4
-            ).split_text(
-                json_data=rendered_doc,
-                convert_lists=True,
-            )
-        ]
 
 
 class LearningResourcesVectorSearchRequestSerializer(serializers.Serializer):
